mask merged with corr module
