--- tomopy/misc/mask.py
@@ -1,123 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-# #########################################################################
-# Copyright (c) 2015, UChicago Argonne, LLC. All rights reserved.         #
-#                                                                         #
-# Copyright 2015. UChicago Argonne, LLC. This software was produced       #
-# under U.S. Government contract DE-AC02-06CH11357 for Argonne National   #
-# Laboratory (ANL), which is operated by UChicago Argonne, LLC for the    #
-# U.S. Department of Energy. The U.S. Government has rights to use,       #
-# reproduce, and distribute this software.  NEITHER THE GOVERNMENT NOR    #
-# UChicago Argonne, LLC MAKES ANY WARRANTY, EXPRESS OR IMPLIED, OR        #
-# ASSUMES ANY LIABILITY FOR THE USE OF THIS SOFTWARE.  If software is     #
-# modified to produce derivative works, such modified software should     #
-# be clearly marked, so as not to confuse it with the version available   #
-# from ANL.                                                               #
-#                                                                         #
-# Additionally, redistribution and use in source and binary forms, with   #
-# or without modification, are permitted provided that the following      #
-# conditions are met:                                                     #
-#                                                                         #
-#     * Redistributions of source code must retain the above copyright    #
-#       notice, this list of conditions and the following disclaimer.     #
-#                                                                         #
-#     * Redistributions in binary form must reproduce the above copyright #
-#       notice, this list of conditions and the following disclaimer in   #
-#       the documentation and/or other materials provided with the        #
-#       distribution.                                                     #
-#                                                                         #
-#     * Neither the name of UChicago Argonne, LLC, Argonne National       #
-#       Laboratory, ANL, the U.S. Government, nor the names of its        #
-#       contributors may be used to endorse or promote products derived   #
-#       from this software without specific prior written permission.     #
-#                                                                         #
-# THIS SOFTWARE IS PROVIDED BY UChicago Argonne, LLC AND CONTRIBUTORS     #
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT       #
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS       #
-# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL UChicago     #
-# Argonne, LLC OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,        #
-# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,    #
-# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;        #
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER        #
-# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT      #
-# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN       #
-# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE         #
-# POSSIBILITY OF SUCH DAMAGE.                                             #
-# #########################################################################
-
-"""
-Module for masking of arrays.
-"""
-
-from __future__ import absolute_import, division, print_function
-
-import numpy as np
-import tomopy.util.dtype as dtype
-import logging
-
-logger = logging.getLogger(__name__)
-
-
-__author__ = "Doga Gursoy"
-__copyright__ = "Copyright (c) 2015, UChicago Argonne, LLC."
-__docformat__ = 'restructuredtext en'
-__all__ = ['circ_mask']
-
-
-def circ_mask(arr, axis, ratio=1, val=0.):
-    """
-    Apply circular mask to a 3D array.
-
-    Parameters
-    ----------
-    arr : ndarray
-        Arbitrary 3D array.
-    axis : int
-        Axis along which mask will be performed.
-    ratio : int, optional
-        Ratio of the mask's diameter in pixels to
-        the smallest edge size along given axis.
-    val : int, optional
-        Value for the masked region.
-
-    Returns
-    -------
-    ndarray
-        Masked array.
-    """
-    arr = dtype.as_float32(arr)
-    _arr = arr.swapaxes(0, axis)
-    dx, dy, dz = _arr.shape
-    mask = _get_mask(dy, dz, ratio)
-    for m in range(dx):
-        _arr[m, ~mask] = val
-    return _arr.swapaxes(0, axis)
-
-
-def _get_mask(dx, dy, ratio):
-    """
-    Calculate 2D boolean circular mask.
-
-    Parameters
-    ----------
-    dx, dy : int
-        Dimensions of the 2D mask.
-
-    ratio : int
-        Ratio of the circle's diameter in pixels to
-        the smallest mask dimension.
-
-    Returns
-    -------
-    ndarray
-        2D boolean array.
-    """
-    rad1 = dx / 2.
-    rad2 = dy / 2.
-    if dx < dy:
-        r2 = rad1 * rad1
-    else:
-        r2 = rad2 * rad2
-    y, x = np.ogrid[0.5 - rad1:0.5 + rad1, 0.5 - rad2:0.5 + rad2]
-    return x * x + y * y < ratio * ratio * r2
